remove people module
